# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Satpol PP",
        "description": "Dokumentasi ini berisi semua endpoint API yang tersedia untuk proyek Satpol PP.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/struktur-organisasi": {
            "get": {
                "tags": [
                    "Struktur Organisasi"
                ],
                "summary": "Menampilkan semua data struktur organisasi",
                "description": "Mengambil semua data yang diurutkan berdasarkan seksi dan urutan, lalu dikelompokkan berdasarkan seksi.",
                "operationId": "getStructures",
                "responses": {
                    "200": {
                        "description": "Operasi berhasil"
                    }
                }
            },
            "post": {
                "tags": [
                    "Struktur Organisasi"
                ],
                "summary": "Membuat data struktur organisasi baru",
                "operationId": "storeStructure",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "position",
                                    "section",
                                    "order"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "Kepala Seksi"
                                    },
                                    "nip": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "photo": {
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "section": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "order": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/struktur-organisasi/{id}": {
            "get": {
                "tags": [
                    "Struktur Organisasi"
                ],
                "summary": "Menampilkan detail satu data struktur organisasi",
                "operationId": "getStructureById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operasi berhasil"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "post": {
                "tags": [
                    "Struktur Organisasi"
                ],
                "summary": "Memperbarui data struktur organisasi (gunakan method POST dengan _method=PUT)",
                "description": "Karena form-data tidak secara native mendukung PUT, kita 'menumpang' di method POST.",
                "operationId": "updateStructure",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "position",
                                    "section",
                                    "order",
                                    "_method"
                                ],
                                "properties": {
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "position": {
                                        "type": "string"
                                    },
                                    "photo": {
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "section": {
                                        "type": "integer"
                                    },
                                    "order": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diperbarui"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Struktur Organisasi"
                ],
                "summary": "Menghapus data struktur organisasi",
                "operationId": "deleteStructure",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil dihapus"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}